requirements.txt
aws-ssm-copy==0.5.2
boto3==1.28.8
botocore==1.31.8
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
